pyproject.toml
[tool.poetry]
name = "drf-pipeline-views"
version = "0.8.0"
description = "Django REST framework views using the pipeline pattern."
authors = [
    "Matti Lamppu <lamppu.matti.akseli@gmail.com>",
]
packages = [
    { include = "pipeline_views" },
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MrThearMan/drf-pipeline-views"
repository = "https://github.com/MrThearMan/drf-pipeline-views"
keywords = [
    "django",
    "djangorestframework",
    "drf",
    "pipeline",
    "views",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Intended Audience :: Developers",
    "Natural Language :: English",
]
include = [
    "README.md",
    "LICENSE",
]
exclude = [
    "tests",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/MrThearMan/drf-pipeline-views/issues"

[tool.poetry.dependencies]
python = ">=3.7.2,<4"
Django = ">=3.0"
djangorestframework = ">=3.12.0"
asgiref = ">=3.5.0"
PyYAML = ">=6.0"
typing-extensions = {version = ">=4.4.0", python = "<3.11"}
pydantic = {version = ">=1.6.2", optional = true}
uvloop = {version = ">=0.16.0", optional = true}

[tool.poetry.group.test.dependencies]
pytest = "7.2.0"
coverage = "6.5.0"
pytest-django = "4.5.2"
pre-commit = "2.20.0"
tox = "3.27.1"
tox-gh-actions = "2.11.0"
coveralls = "3.3.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "1.4.2"
pymdown-extensions = "9.8"
mkdocs-material = "8.5.10"
Pygments = "2.13.0"

[tool.poetry.group.lint.dependencies]
mypy = "0.991"
django-stubs = "1.13.0"
djangorestframework-stubs = "1.8.0"

[tool.poetry.extras]
uvloop = ["uvloop"]
pydantic = ["pydantic"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = "True"
lines_after_imports = 2
known_third_party = [
    "django",
    "rest_framework",
]

[tool.ruff]
fix = true
line-length = 120
exclude = [
    "tests/*",
]
select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    # "D",  # pydocstyle
    # "N",  # pep8-naming
    "S",  # flake8-bandit
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "T",  # flake8-print
    "W",  # pycodestyle warnings
    # "ANN",  # flake8-annotations
]
ignore = [
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"pipeline_views/schema.py" = ["C901"]


[tool.mypy]
python_version = "3.10"
warn_return_any = "True"
warn_unused_configs = "True"
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[tool.django-stubs]
django_settings_module = "tests.django.settings"

[tool.coverage.run]
relative_files = true

[tool.coverage.report]
omit = [
    "tests/*",
    "docs/*",
    ".venv/*",
    ".tox/*",
]
exclude_lines = [
    "if TYPE_CHECKING:",
    "except ImportError:",
    "pragma: no cover",
]

[tool.pytest.ini_options]
addopts = "-vv -s --log-cli-level=INFO"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py{37, 38, 39, 310, 311}-drf{312, 313, 314}
isolated_build = true

[gh-actions]
python =
    3.7: py37
    3.8: py38
    3.9: py39
    3.10: py310
    3.11: py311

[testenv]
whitelist_externals =
    poetry
setenv =
    PYTHONPATH = {toxinidir}
    DJANGO_SETTINGS_MODULE = tests.django.settings
deps =
    py37: Django >=3.0,<3.1
    py38: Django >=3.1,<3.2
    py39: Django >=3.2,<4.0

    py310-drf312: Django >=3.2,<4.0
    py310-drf313: Django >=4.0,<4.1
    py310-drf314: Django >=4.1,<4.2

    py311-drf312: Django >=3.2,<4.0
    py311-drf313: Django >=4.0,<4.1
    py311-drf314: Django >=4.1,<4.2

    drf312: djangorestframework >=3.12.0,<3.13.0
    drf313: djangorestframework >=3.13.0,<3.14.0
    drf314: djangorestframework >=3.14.0,<3.15.0

    pytest
    coverage
    pytest-django
    asgiref
    pydantic
    pyyaml
commands =
    coverage run -m pytest -vv -s --log-cli-level=INFO
"""

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
